Move to configmodel
